+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.13
+
+- Add possibility to move and resize areas
+- Remove `rows` configuration cause it is now dynamic depending on the areas you create
+- Add possibility to delete content/area by clicking on it with the `meta` key pressed
+
 ## 0.0.12
 
 - Add `width`, `height`, `thumbhash`, `alt` and `title` property in returned content object